+++ steady_pond.h
@@ -61,9 +61,9 @@ public:
     }
 
     template <typename _Container>
-    void enqueue(_Container& cont, uint size) {
+    void enqueue(_Container& cont, size_t size) {
         util::spinlock_guard lock(tq_locker);
-        for (int i = 0; i < size; ++i) {
+        for (size_t i = 0; i < size; ++i) {
             public_tq.emplace(std::move(cont[i]));
             task_numb++;
         }
@@ -85,7 +85,7 @@ public:
      * @param thread_numb fixed thread number
      * @param task_capacity task capacity of the pond, default: unlimited
     */
-    SteadyThreadPond(uint thread_numb = 0, uint task_capacity = HipeUnlimited) 
+    SteadyThreadPond(int thread_numb = 0, int task_capacity = HipeUnlimited) 
         : FixedThreadPond(thread_numb, task_capacity)
     {
         // create threads
